refactor: add component directories, move css files to component directories, change imports
