+++ src/App.tsx
@@ -1,9 +1,9 @@
 import './App.css';
-import SearchSection from './components/SearchSection';
-import ErrorBoundary from './components/ErrorBoundary';
+import { SearchSection } from './components/SearchSection';
+import { ErrorBoundary } from './components/ErrorBoundary';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Details from './components/Details';
-import NotFoundPage from './components/NotFoundPage';
+import { Details } from './components/Details';
+import { NotFoundPage} from './components/NotFoundPage';
 
 export default function App() {
   return (
