+++ src/__ tests __/NotFoundPage.test.tsx
@@ -2,9 +2,9 @@ import '@testing-library/jest-dom';
 import { render, screen } from '@testing-library/react';
 import { Route, Routes } from 'react-router-dom';
 import { MemoryRouter } from 'react-router-dom';
-import Details from '../components/Details';
-import NotFoundPage from '../components/NotFoundPage';
-import SearchSection from '../components/SearchSection';
+import {Details} from '../components/Details';
+import {NotFoundPage} from '../components/NotFoundPage';
+import {SearchSection} from '../components/SearchSection';
 
 describe('PageNotFound', () => {
   test('404 page is displayed when navigating to an invalid route', () => {
