+++ src/components/CardList/index.ts
@@ -1,0 +1,1 @@
+export { CardsList } from "./CardsList";
