+++ src/components/Details/Details.css
