+++ src/components/Details/index.ts
@@ -1,0 +1,1 @@
+export { Details} from "./Details";
