+++ src/components/ErrorButton/index.ts
@@ -1,0 +1,1 @@
+export { ErrorButton }  from "./ErrorButton";
