+++ src/components/Loader/Loader.tsx
@@ -1,6 +1,6 @@
-import './styles/Loader.css';
+import './Loader.css';
 
-export default function Loader() {
+export function Loader() {
   return (
     <div className="loader-wrapper" data-testid="loader">
       <img className="loader" src="./svg/pokeball.svg" alt="loader"></img>;
