+++ src/components/SearchSection/SearchSection.css
