+++ src/components/SearchSection/SearchSection.tsx
@@ -1,23 +1,23 @@
 import { useEffect, useState } from 'react';
-import CardsList from './CardsList';
-import { CardData } from '../types/interfaces';
-import SearchBar from './SearchBar';
-import Loader from './Loader';
-import useLocalStorage from './hooks/useLocalStorage';
-import Pagination from './Pagination';
+import { CardsList } from '../CardList';
+import { CardData } from '../../types/interfaces';
+import { SearchBar } from '../SearchBar';
+import { Loader } from '../Loader';
+import useLocalStorage from '../../hooks/useLocalStorage';
+import { Pagination } from '../Pagination';
 import {
   Outlet,
   useNavigate,
   useLocation,
   useSearchParams,
 } from 'react-router-dom';
-import './styles/SearchSection.css';
-import { searchCardsByName } from '../Api';
+import { searchCardsByName } from '../../Api';
+import './SearchSection.css';
 
 const lsQueryKey = 'searchQuery';
 const lsPageSizeKey = 'cardsPerPage';
 
-export default function SearchSection() {
+export function SearchSection() {
   const [lsQueryValue, setLsQueryValue] = useLocalStorage(lsQueryKey);
   const [lsPageSizeValue, setPageSizeValue] = useLocalStorage(lsPageSizeKey);
